feat: ismodel 일 경우 분기처리
